chore(lambda): refactor utils funcs.
